requirements.txt
automium==0.2.6
automium-web==0.1.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
dj-database-url==0.4.2
Django==1.11.2
django-filter==1.1.0
django-mathfilters==0.4.0
django-taggit==0.22.2
djangorestframework==3.7.1
Flask==0.12.2
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
legacy==0.1.6
Markdown==2.6.9
MarkupSafe==1.0
Pillow==5.0.0
post==0.0.13
psycopg2==2.7.1
PyPDF2==1.26.0
pytz==2017.2
reportlab==3.4.0
requests==2.18.4
selenium==3.7.0
setupfiles==0.0.50
simplejson==3.11.1
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==3.3.0
xhtml2pdf==0.2b1
xlrd==1.1.0
yahoo-finance==1.4.0
